requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.3
django-cors==0.1
django-cors-headers==3.13.0
django-import-export==3.0.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.3.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.3
tablib==3.2.1
tzdata==2022.7
urllib3==1.26.13
xlrd==2.0.1
xlwt==1.3.0